+++ src/main/java/roomescape/admin/reservation/controller/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.admin.reservation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
