+++ src/main/java/roomescape/admin/reservation/dto/response/ReservationResponse.java
@@ -1,7 +1,8 @@
-package roomescape.admin.reservation;
+package roomescape.admin.reservation.dto.response;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.admin.reservation.entity.Reservation;
 
 public record ReservationResponse(Long id, String name, LocalDate date, LocalTime time) {
     public static ReservationResponse from(Reservation reservation) {
