+++ src/main/java/roomescape/admin/reservation/dto/response/ReservationTimeResponse.java
@@ -1,6 +1,7 @@
-package roomescape.admin.reservation.time;
+package roomescape.admin.reservation.dto.response;
 
 import java.time.LocalTime;
+import roomescape.admin.reservation.entity.ReservationTime;
 
 public record ReservationTimeResponse(Long id, LocalTime startAt) {
     public static ReservationTimeResponse from(ReservationTime reservationTime) {
