+++ src/main/java/roomescape/admin/reservation/entity/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.admin.reservation.time;
+package roomescape.admin.reservation.entity;
 
 import java.time.LocalTime;
