+++ src/main/java/roomescape/admin/reservation/repository/H2ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.admin.reservation;
+package roomescape.admin.reservation.repository;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,44 +8,43 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.admin.reservation.entity.ReservationTime;
 
 @Repository
-public class H2ReservationRepository implements ReservationRepository {
+public class H2ReservationTimeRepository implements ReservationTimeRepository {
 
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
     @Autowired
-    public H2ReservationRepository(JdbcTemplate jdbcTemplate) {
+    public H2ReservationTimeRepository(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
         this.jdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("reservation")
+                .withTableName("reservation_time")
                 .usingGeneratedKeyColumns("id");
     }
 
     @Override
-    public List<Reservation> findAll() {
-        String query = "SELECT id, name, date, time FROM reservation";
-        RowMapper<Reservation> reservationRowMapper = (rs, rowNum) -> new Reservation(
+    public List<ReservationTime> findAll() {
+        String query = "SELECT id, start_at FROM reservation_time";
+        RowMapper<ReservationTime> timeRowMapper = (rs, rowNum) -> new ReservationTime(
                 rs.getLong("id"),
-                rs.getString("name"),
-                rs.getDate("date").toLocalDate(),
-                rs.getTime("time").toLocalTime()
+                rs.getTime("start_at").toLocalTime()
         );
-        return jdbcTemplate.query(query, reservationRowMapper);
+        return jdbcTemplate.query(query, timeRowMapper);
     }
 
     @Override
-    public Reservation save(Reservation reservation) {
+    public ReservationTime save(ReservationTime reservation) {
         SqlParameterSource params = new BeanPropertySqlParameterSource(reservation);
         Long id = jdbcInsert.executeAndReturnKey(params).longValue();
 
-        return new Reservation(id, reservation.getName(), reservation.getDate(), reservation.getTime());
+        return new ReservationTime(id, reservation.getStartAt());
     }
 
     @Override
     public int delete(Long id) {
-        String sql = "DELETE FROM reservation WHERE id = ?";
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
         return jdbcTemplate.update(sql, id);
     }
 }
