+++ src/main/java/roomescape/admin/reservation/repository/ReservationRepository.java
@@ -1,6 +1,7 @@
-package roomescape.admin.reservation;
+package roomescape.admin.reservation.repository;
 
 import java.util.List;
+import roomescape.admin.reservation.entity.Reservation;
 
 public interface ReservationRepository {
